Separate files for NodeRowObserver's cached vs non-cached properties
