+++ stitch.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "f6d92aa9e120daffeadab0dc44ebdf42222187a6f4ed96b4dd6115efc34bfef2",
+  "originHash" : "fa62414f3ffde63efcf96a3e2f8f06cdb8551ff9a33d61d5f1fa2ad8c3db93d0",
   "pins" : [
     {
       "identity" : "audiokit",
